contactlist/requirements.txt

Babel==2.14.0
crispy-bootstrap5==2023.10
Django==4.2.10
django-bootstrap-v5==1.0.11
django-bootstrap5==
django-crispy-forms==2.1
django-phonenumber-field==7.3.0
mysqlclient==2.2.4
phonenumbers==8.13.30
pillow==10.2.0